projects/flask_jwt_auth/requirements.txt
Flask==2.2.3
Werkzeug==2.2.2
flask-jwt-extended==4.2.1
PyJWT==2.11.0
requests==2.31.0
colorama==0.4.6
passlib==1.7.4
